bbb-vpn: adds a hack to filter out all HNAs for default-routes
This hack removes all default-routes from the incoming HNA-
messages. This way there will be only the intra-mesh-routes
avail and mesh-coulds can connect to each other without
sending their default-routes into the mesh.

For the BBB-VPN of Berlin this ensures that the mesh-clouds
will not push their Uplink into the mesh and also will not
receive the uplink-gateways of the mesh.

This hack was running serveral years on the server.y